+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.31
+* Option `showErrorsOnUpdate` renamed `showErrorsOnChange`.
+* Option `retrieveAlsoDisabledFields` renamed `retrieveOnlyEnabledFields`.
+* Option `retrieveOnlyEnabledFields` is `false` by default.
+* `option` Field prop renamed `extra`.
+* `options` Field prop handle individual **Field Options**.
+* **Field Options** will fallback on **Form Options**.
+* Implemented input **file** `type` handling.
+* Introduced `onDrop` Field Event Handler.
+* Introduced `onDrop` Field Hook.
+* Introduced `files` Field prop.
+* Dropzone support.
+
 # 1.30
 * New `onClear` & `onReset` Form & Fields Hooks.
 * Introduced `retrieveAlsoDisabledFields` option.